+++ docs/releases.md
@@ -14,6 +14,7 @@ Patch:
 * Documentation improved in synchronizer
 * Support of extension added in `createTargetsLibrary`
 * Fixing dependency to `aepp` `som` module
+* Fixing renaming path to use `renameComponent` and using `syncComponent`
 
 ## 0.4.4
 * adding a `getBuilds` method on the `Library` class instance